factors out getUsername and makes requests optional
