+++ projects/github-actions/required-review/README.md
@@ -65,6 +65,10 @@ This action is intended to be triggered by the `pull_request_review` event.
     # this to instead fail the status checks instead of leaving them pending.
     fail: true
 
+    # By default required reviewers are not requested. Set this to true to 
+    # request reviews.
+    request-reviews: true
+
     # GitHub Access Token. The user associated with this token will show up
     # as the "creator" of the status check, and must have access to read
     # pull request data, create status checks (`repo:status`), and to read
