+++ projects/github-actions/required-review/action.yml
@@ -23,6 +23,12 @@ inputs:
     required: false
     type: boolean
     default: false
+  request-reviews:
+    description: >
+      Requests reviews from teams or users needed to fulfill a requirement.
+    required: false
+    type: boolean
+    default: false
   token:
     description: >
       GitHub Access Token. The user associated with this token will show up
